refactor(metadata): Refactor Cargo module into submodules.
